LaunchDarkly.Server

Description

This module re-exports the User, Client, and Config modules.

Documentation

data Config

Config allows advanced configuration of the LaunchDarkly client.

makeConfig :: Text -> Config

Create a default configuration from a given SDK key.

configSetKey :: Text -> Config -> Config

Set the SDK key used to authenticate with LaunchDarkly.

configSetBaseURI :: Text -> Config -> Config

The base URI of the main LaunchDarkly service. This should not normally be changed except for testing.

configSetStreamURI :: Text -> Config -> Config

The base URI of the LaunchDarkly streaming service. This should not normally be changed except for testing.

configSetEventsURI :: Text -> Config -> Config

The base URI of the LaunchDarkly service that accepts analytics events. This should not normally be changed except for testing.

configSetStreaming :: Bool -> Config -> Config

Sets whether streaming mode should be enabled. By default, streaming is enabled. It should only be disabled on the advice of LaunchDarkly support.

configSetAllAttributesPrivate :: Bool -> Config -> Config

Sets whether or not all user attributes (other than the key) should be hidden from LaunchDarkly. If this is true, all user attribute values will be private, not just the attributes specified in PrivateAttributeNames.

configSetPrivateAttributeNames :: Set Text -> Config -> Config

Marks a set of user attribute names private. Any users sent to LaunchDarkly with this configuration active will have attributes with these names removed.

configSetFlushIntervalSeconds :: Natural -> Config -> Config

The time between flushes of the event buffer. Decreasing the flush interval means that the event buffer is less likely to reach capacity.

configSetPollIntervalSeconds :: Natural -> Config -> Config

The polling interval (when streaming is disabled).

configSetUserKeyLRUCapacity :: Natural -> Config -> Config

The number of user keys that the event processor can remember at any one time, so that duplicate user details will not be sent in analytics events.

configSetInlineUsersInEvents :: Bool -> Config -> Config

Set to true if you need to see the full user details in every analytics event.

configSetEventsCapacity :: Natural -> Config -> Config

The capacity of the events buffer. The client buffers up to this many events in memory before flushing. If the capacity is exceeded before the buffer is flushed, events will be discarded.

configSetLogger :: (LoggingT IO () -> IO ()) -> Config -> Config

Set the logger to be used by the client.

configSetSendEvents :: Bool -> Config -> Config

Sets whether to send analytics events back to LaunchDarkly. By default, the client will send events. This differs from Offline in that it only affects sending events, not streaming or polling for events from the server.

configSetOffline :: Bool -> Config -> Config

Sets whether this client is offline. An offline client will not make any network connections to LaunchDarkly, and will return default values for all feature flags.

configSetRequestTimeoutSeconds :: Natural -> Config -> Config

Sets how long an the HTTP client should wait before a response is returned.

configSetStoreBackend :: Maybe StoreInterface -> Config -> Config

Configures a handle to an external store such as Redis.

configSetStoreTTL :: Natural -> Config -> Config

When a store backend is configured, control how long values should be cached in memory before going back to the backend.

configSetUseLdd :: Bool -> Config -> Config

Sets whether this client should use the LaunchDarkly relay in daemon mode. In this mode, the client does not subscribe to the streaming or polling API, but reads data only from the feature store. See: https://docs.launchdarkly.com/docs/the-relay-proxy

data User

User contains specific attributes of a user of your application

The only mandatory property is the Key, which must uniquely identify each user. For authenticated users, this may be a username or e-mail address. For anonymous users, this could be an IP address or session ID.

makeUser :: Text -> User

Creates a new user identified by the given key.

userSetKey :: Maybe Text -> User -> User

Set the primary key for a user.

userSetSecondary :: Maybe Text -> User -> User

Set the secondary key for a user.

userSetIP :: Maybe Text -> User -> User

Set the IP for a user.

userSetCountry :: Maybe Text -> User -> User

Set the country for a user.

userSetEmail :: Maybe Text -> User -> User

Set the email for a user.

userSetFirstName :: Maybe Text -> User -> User

Set the first name for a user.

userSetLastName :: Maybe Text -> User -> User

Set the last name for a user.

userSetAvatar :: Maybe Text -> User -> User

Set the avatar for a user.

userSetName :: Maybe Text -> User -> User

Set the name for a user.

userSetAnonymous :: Bool -> User -> User

Set if the user is anonymous or not.

userSetCustom :: HashMap Text Value -> User -> User

Set custom fields for a user.

userSetPrivateAttributeNames :: Set Text -> User -> User

This contains list of attributes to keep private, whether they appear at the top-level or Custom The attribute "key" is always sent regardless of whether it is in this list, and "custom" cannot be used to eliminate all custom attributes

data Client

Client is the LaunchDarkly client. Client instances are thread-safe. Applications should instantiate a single instance for the lifetime of their application.

makeClient :: Config -> IO Client

Create a new instance of the LaunchDarkly client.

clientVersion :: Text

The version string for this library.

boolVariation :: Client -> Text -> User -> Bool -> IO Bool

Evaluate a Boolean typed flag.

boolVariationDetail :: Client -> Text -> User -> Bool -> IO (EvaluationDetail Bool)

Evaluate a Boolean typed flag, and return an explation.

stringVariation :: Client -> Text -> User -> Text -> IO Text

Evaluate a String typed flag.

stringVariationDetail :: Client -> Text -> User -> Text -> IO (EvaluationDetail Text)

Evaluate a String typed flag, and return an explanation.

intVariation :: Client -> Text -> User -> Int -> IO Int

Evaluate a Number typed flag, and truncate the result.

intVariationDetail :: Client -> Text -> User -> Int -> IO (EvaluationDetail Int)

Evaluate a Number typed flag, truncate the result, and return an explanation.

doubleVariation :: Client -> Text -> User -> Double -> IO Double

Evaluate a Number typed flag.

doubleVariationDetail :: Client -> Text -> User -> Double -> IO (EvaluationDetail Double)

Evaluate a Number typed flag, and return an explanation.

jsonVariation :: Client -> Text -> User -> Value -> IO Value

Evaluate a JSON typed flag.

jsonVariationDetail :: Client -> Text -> User -> Value -> IO (EvaluationDetail Value)

Evaluate a JSON typed flag, and return an explanation.

data EvaluationDetail value

Combines the result of a flag evaluation with an explanation of how it was calculated.

Constructors

EvaluationDetail

Fields

value :: !value

The result of the flag evaluation. This will be either one of the flag's variations or the default value passed by the application.

variationIndex :: !(Maybe Natural)

The index of the returned value within the flag's list of variations, e.g. 0 for the first variation - or Nothing if the default value was returned.

reason :: !EvaluationReason

Describes the main factor that influenced the flag evaluation value.

data EvaluationReason

Defines the possible values of the Kind property of EvaluationReason.

Constructors

EvaluationReasonOff
Indicates that the flag was off and therefore returned its configured off value.

EvaluationReasonTargetMatch
indicates that the user key was specifically targeted for this flag.

EvaluationReasonRuleMatch
Indicates that the user matched one of the flag's rules.

Fields

ruleIndex :: !Natural

The index of the rule that was matched (0 being the first).

ruleId :: !Text

The unique identifier of the rule that was matched.

EvaluationReasonPrerequisiteFailed
Indicates that the flag was considered off because it had at least one prerequisite flag that either was off or did not return the desired variation.

Fields

prerequisiteKey :: !Text

The flag key of the prerequisite that failed.

EvaluationReasonFallthrough
Indicates that the flag was on but the user did not match any targets or rules.

EvaluationReasonError
Indicates that the flag could not be evaluated, e.g. because it does not exist or due to an unexpected error. In this case the result value will be the default value that the caller passed to the client.

Fields

errorKind :: !EvalErrorKind

Describes the type of error.

data EvalErrorKind

Defines the possible values of the errorKind property of EvaluationReason.

Constructors

EvalErrorKindMalformedFlag
Indicates that there was an internal inconsistency in the flag data, e.g. a rule specified a nonexistent variation.

EvalErrorFlagNotFound
Indicates that the caller provided a flag key that did not match any known flag.

EvalErrorWrongType
Indicates that the result value was not of the requested type, e.g. you called boolVariationDetail but the value was an integer.

EvalErrorUserNotSpecified
Indicates that the caller passed a user without a key for the user parameter.

EvalErrorClientNotReady
Indicates that the caller tried to evaluate a flag before the client had successfully initialized.

EvalErrorExternalStore !Text
Indicates that some error was returned by the external feature store.

allFlags :: Client -> User -> IO (HashMap Text Value)

Returns a map from feature flag keys to values for a given user. If the result of the flag's evaluation would result in the default value, Null will be returned. This method does not send analytics events back to LaunchDarkly.

close :: Client -> IO ()

Close shuts down the LaunchDarkly client. After calling this, the LaunchDarkly client should no longer be used. The method will block until all pending analytics events have been sent.

flushEvents :: Client -> IO ()

Flush tells the client that all pending analytics events (if any) should be delivered as soon as possible. Flushing is asynchronous, so this method will return before it is complete.

identify :: Client -> User -> IO ()

Identify reports details about a a user.

track :: Client -> User -> Text -> Maybe Value -> Maybe Double -> IO ()

Track reports that a user has performed an event. Custom data can be attached to the event, and / or a numeric value.

The numeric value is used by the LaunchDarkly experimentation feature in numeric custom metrics, and will also be returned as part of the custom event for Data Export.

data Status

The status of the client initialization.

Constructors

Uninitialized
The client has not yet finished connecting to LaunchDarkly.

Unauthorized
The client attempted to connect to LaunchDarkly and was denied.

Initialized
The client has successfuly connected to LaunchDarkly.

ShuttingDown
The client is being terminated

getStatus :: Client -> IO Status

Return the initialization status of the Client