LaunchDarkly.Server.Client

Description

This module contains the core functionality of the SDK.

Documentation

data Client

Client is the LaunchDarkly client. Client instances are thread-safe. Applications should instantiate a single instance for the lifetime of their application.

makeClient :: Config -> IO Client

Create a new instance of the LaunchDarkly client.

clientVersion :: Text

The version string for this library.

boolVariation :: Client -> Text -> Context -> Bool -> IO Bool

Evaluate a Boolean typed flag.

boolVariationDetail :: Client -> Text -> Context -> Bool -> IO (EvaluationDetail Bool)

Evaluate a Boolean typed flag, and return an explanation.

stringVariation :: Client -> Text -> Context -> Text -> IO Text

Evaluate a String typed flag.

stringVariationDetail :: Client -> Text -> Context -> Text -> IO (EvaluationDetail Text)

Evaluate a String typed flag, and return an explanation.

intVariation :: Client -> Text -> Context -> Int -> IO Int

Evaluate a Number typed flag, and truncate the result.

intVariationDetail :: Client -> Text -> Context -> Int -> IO (EvaluationDetail Int)

Evaluate a Number typed flag, truncate the result, and return an explanation.

doubleVariation :: Client -> Text -> Context -> Double -> IO Double

Evaluate a Number typed flag.

doubleVariationDetail :: Client -> Text -> Context -> Double -> IO (EvaluationDetail Double)

Evaluate a Number typed flag, and return an explanation.

jsonVariation :: Client -> Text -> Context -> Value -> IO Value

Evaluate a JSON typed flag.

jsonVariationDetail :: Client -> Text -> Context -> Value -> IO (EvaluationDetail Value)

Evaluate a JSON typed flag, and return an explanation.

data EvaluationDetail value

Combines the result of a flag evaluation with an explanation of how it was calculated.

Constructors

EvaluationDetail

Fields

value :: !value

The result of the flag evaluation. This will be either one of the flag's variations or the default value passed by the application.

variationIndex :: !(Maybe Integer)

The index of the returned value within the flag's list of variations, e.g. 0 for the first variation - or Nothing if the default value was returned.

reason :: !EvaluationReason

Describes the main factor that influenced the flag evaluation value.

data EvaluationReason

Defines the possible values of the Kind property of EvaluationReason.

Constructors

EvaluationReasonOff
Indicates that the flag was off and therefore returned its configured off value.

EvaluationReasonTargetMatch
indicates that the context key was specifically targeted for this flag.

EvaluationReasonRuleMatch

Fields

ruleIndex :: !Natural

The index of the rule that was matched (0 being the first).

ruleId :: !Text

The unique identifier of the rule that was matched.

inExperiment :: !Bool

Whether the evaluation was part of an experiment. Is true if the evaluation resulted in an experiment rollout *and* served one of the variations in the experiment. Otherwise false.

EvaluationReasonPrerequisiteFailed

Fields

prerequisiteKey :: !Text

The flag key of the prerequisite that failed.

EvaluationReasonFallthrough

Fields

inExperiment :: !Bool

Whether the evaluation was part of an experiment. Is true if the evaluation resulted in an experiment rollout *and* served one of the variations in the experiment. Otherwise false.

EvaluationReasonError

Fields

errorKind :: !EvalErrorKind

Describes the type of error.

data EvalErrorKind

Defines the possible values of the errorKind property of EvaluationReason.

Constructors

EvalErrorKindMalformedFlag
Indicates that there was an internal inconsistency in the flag data, e.g. a rule specified a nonexistent variation.

EvalErrorFlagNotFound
Indicates that the caller provided a flag key that did not match any known flag.

EvalErrorWrongType
Indicates that the result value was not of the requested type, e.g. you called boolVariationDetail but the value was an integer.

EvalErrorClientNotReady
Indicates that the caller tried to evaluate a flag before the client had successfully initialized.

EvalErrorInvalidContext
Indicates that the caller tried to evaluate a flag with an invalid context

EvalErrorExternalStore !Text
Indicates that some error was returned by the external feature store.

allFlagsState :: Client -> Context -> Bool -> Bool -> Bool -> IO AllFlagsState

Returns an object that encapsulates the state of all feature flags for a given context. This includes the flag values, and also metadata that can be used on the front end.

The most common use case for this method is to bootstrap a set of client-side feature flags from a back-end service.

The first parameter will limit to only flags that are marked for use with the client-side SDK (by default, all flags are included).

The second parameter will include evaluation reasons in the state.

The third parameter will omit any metadata that is normally only used for event generation, such as flag versions and evaluation reasons, unless the flag has event tracking or debugging turned on

For more information, see the Reference Guide: https://docs.launchdarkly.com/sdk/features/all-flags#haskell

data AllFlagsState

AllFlagsState captures the state of all feature flag keys as evaluated for a specific context. This includes their values, as well as other metadata.

secureModeHash :: Client -> Context -> Text

Generates the secure mode hash value for a context.

For more information, see the Reference Guide: https://docs.launchdarkly.com/sdk/features/secure-mode#haskell.

close :: Client -> IO ()

Close shuts down the LaunchDarkly client. After calling this, the LaunchDarkly client should no longer be used. The method will block until all pending analytics events have been sent.

flushEvents :: Client -> IO ()

Flush tells the client that all pending analytics events (if any) should be delivered as soon as possible. Flushing is asynchronous, so this method will return before it is complete.

identify :: Client -> Context -> IO ()

Identify reports details about a context.

track :: Client -> Context -> Text -> Maybe Value -> Maybe Double -> IO ()

Track reports that a context has performed an event. Custom data can be attached to the event, and / or a numeric value.

The numeric value is used by the LaunchDarkly experimentation feature in numeric custom metrics, and will also be returned as part of the custom event for Data Export.

data Status

The status of the client initialization.

Constructors

Uninitialized
The client has not yet finished connecting to LaunchDarkly.

Unauthorized
The client attempted to connect to LaunchDarkly and was denied.

Initialized
The client has successfuly connected to LaunchDarkly.

ShuttingDown
The client is being terminated

getStatus :: Client -> IO Status

Return the initialization status of the Client