LaunchDarkly.Server.Store

Description

This module contains details for external store implementations.

Documentation

type StoreResult a = IO (Either Text a)

The result type for every StoreInterface function. Instead of throwing an exception, any store related error should return Left. Exceptions unrelated to the store should not be caught.

type FeatureKey = Text

Represents the key for a given feature.

type FeatureNamespace = Text

Represents a namespace such as flags or segments

data StoreInterface

The interface implemented by external stores for use by the SDK.

Constructors

StoreInterface

Fields

storeInterfaceAllFeatures :: !(FeatureNamespace -> StoreResult (HashMap Text RawFeature))

A map of all features in a given namespace including deleted.

storeInterfaceGetFeature :: !(FeatureNamespace -> FeatureKey -> StoreResult RawFeature)

Return the value of a key in a namespace.

storeInterfaceUpsertFeature :: !(FeatureNamespace -> FeatureKey -> RawFeature -> StoreResult Bool)

Upsert a given feature. Versions should be compared before upsert. The result should indicate if the feature was replaced or not.

storeInterfaceIsInitialized :: !(StoreResult Bool)

Checks if the external store has been initialized, which may have been done by another instance of the SDK.

storeInterfaceInitialize :: !(HashMap FeatureNamespace (HashMap FeatureKey RawFeature) -> StoreResult ())

A map of namespaces, and items in namespaces. The entire store state should be replaced with these values.

data RawFeature

An abstract representation of a store object.

Constructors

RawFeature

Fields

rawFeatureBuffer :: !(Maybe ByteString)

A serialized item. If the item is deleted or does not exist this should be Nothing.

rawFeatureVersion :: !Natural

The version of a given item. If the item does not exist this should be zero.