LaunchDarkly.Server.User

Description

This module is for configuration of the user object.

Documentation

data User

User contains specific attributes of a user of your application

The only mandatory property is the Key, which must uniquely identify each user. For authenticated users, this may be a username or e-mail address. For anonymous users, this could be an IP address or session ID.

makeUser :: Text -> User

Creates a new user identified by the given key.

userSetKey :: Text -> User -> User

Set the primary key for a user.

userSetSecondary :: Maybe Text -> User -> User

Set the secondary key for a user.

userSetIP :: Maybe Text -> User -> User

Set the IP for a user.

userSetCountry :: Maybe Text -> User -> User

Set the country for a user.

userSetEmail :: Maybe Text -> User -> User

Set the email for a user.

userSetFirstName :: Maybe Text -> User -> User

Set the first name for a user.

userSetLastName :: Maybe Text -> User -> User

Set the last name for a user.

userSetAvatar :: Maybe Text -> User -> User

Set the avatar for a user.

userSetName :: Maybe Text -> User -> User

Set the name for a user.

userSetAnonymous :: Bool -> User -> User

Set if the user is anonymous or not.

userSetCustom :: HashMap Text Value -> User -> User

Set custom fields for a user.

userSetPrivateAttributeNames :: Set Text -> User -> User

This contains list of attributes to keep private, whether they appear at the top-level or Custom The attribute "key" is always sent regardless of whether it is in this list, and "custom" cannot be used to eliminate all custom attributes