LaunchDarkly.Server.Store

Description

This module contains details for external store implementations.

Documentation

type StoreResult a = IO (Either Text a)

The result type for every PersistentDataStore function. Instead of throwing an exception, any store related error should return Left. Exceptions unrelated to the store should not be caught.

type FeatureKey = Text

Represents the key for a given feature.

type FeatureNamespace = Text

Represents a namespace such as features or segments

data PersistentDataStore

The interface implemented by external stores for use by the SDK.

Constructors

PersistentDataStore

Fields

persistentDataStoreAllFeatures :: !(FeatureNamespace -> StoreResult (KeyMap SerializedItemDescriptor))

A map of all features in a given namespace including deleted.

persistentDataStoreGetFeature :: !(FeatureNamespace -> FeatureKey -> StoreResult (Maybe SerializedItemDescriptor))

Return the value of a key in a namespace.

persistentDataStoreUpsertFeature :: !(FeatureNamespace -> FeatureKey -> SerializedItemDescriptor -> StoreResult Bool)

Upsert a given feature. Versions should be compared before upsert. The result should indicate if the feature was replaced or not.

persistentDataStoreIsInitialized :: !(StoreResult Bool)

Checks if the external store has been initialized, which may have been done by another instance of the SDK.

persistentDataStoreInitialize :: !(KeyMap (KeyMap SerializedItemDescriptor) -> StoreResult ())

A map of namespaces, and items in namespaces. The entire store state should be replaced with these values.

data SerializedItemDescriptor

A record representing an object that can be persisted in an external store.

Constructors

SerializedItemDescriptor

Fields

item :: !(Maybe ByteString)

A serialized item. If the item is deleted or does not exist this should be Nothing.

version :: !Natural

The version of a given item. If the item does not exist this should be zero.

deleted :: !Bool

True if this is a placeholder (tombstone) for a deleted item.

serializeWithPlaceholder :: SerializedItemDescriptor -> ByteString

Generate a ByteString representation of the SerializedItemDescriptor.

If the SerializedItemDescriptor has either a Nothing value, or is marked as deleted, the ByteString representation will be a tombstone marker containing the version and deletion status.

Otherwise, the internal item representation is returned.

byteStringToVersionedData :: ByteString -> Maybe VersionedData

Partially decode the provided ByteString into a VersionedData struct.

This is useful for persistent stores who need to perform version comparsions before persisting data.